requirements.txt
markitdown[all]>=0.0.1
docling>=2.0.0
python-pptx>=1.0.0
